# Optimize performance reduce round() on double to int