Speedup encoder, remove round() on double.

For cmd line below:
```
-i 1920x1080_60fps_420_8bit.yuv -enc-mode [mode] -w 1920 -h 1080 -bit-depth 8 -fps 60 -b out.ivf
```

Encoding speedup:
 Mode | master[fps] | Changes[fps] | speedup[%]
-- | -- | --| --
M8 | 18,953 | 19,659|3.73
M7 | 11,672 | 11,986|2.69
M6 | 10,454 | 10,761|2.94
M5 | 5,615 | 5,669|0.96
M4 | 3,745 | 3,786|1.09
M3 | 2,483 | 2,511|1.13
M2 | 0,32 | 0,321|0.31
M1 | 0,051 | 0,052|1.96
M0 | 0,037 | 0,037|0.00