@tianjunwork 2020-05-11
Thanks for this optimization! Surprised to see data type casting bring that much speed penalty. Curious, is that because `float` is used? or just casting return value from double to integer.