@spawlows 2020-05-12
> Thanks for this optimization! Surprised to see data type casting bring that much speed penalty. Curious, is that because `float` is used? or just casting return value from double to integer.

Float is one of thing, but this change remove calls function round(), just doing some trick on integers, and one division enough to get that same result.