@EricLamVCD 2020-05-12
Obj-1 8bit and1 10bit clips: 
  | Speed Deviation  | BD-Rate   Deviation
-- | -- | --
M0 - 8bit | 0.80% | 0.00%
M4 - 8bit | 0.70% | 0.00%
M8 - 8bit | 2.40% | 0.00%
M0 - 10bit | 0.52% | 0.00%
M4 - 10bit | -0.15% | 0.00%
M8 - 10bit | 1.21% | 0.00%